+++ z/btree.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 	"unsafe"
 
-	"github.com/dgraph-io/ristretto/contrib/simd"
+	"github.com/dgraph-io/ristretto/z/simd"
 )
 
 var (
@@ -485,6 +485,14 @@ func (n node) isFull() bool {
 // Search returns the index of a smallest key >= k in a node.
 func (n node) search(k uint64) int {
 	N := n.numKeys()
+	if N < 4 {
+		for i := 0; i < N; i++ {
+			if ki := n.key(i); ki >= k {
+				return i
+			}
+		}
+		return N
+	}
 	return int(simd.Search(n[:2*N], k))
 	// lo, hi := 0, N
 	// // Reduce the search space using binary seach and then do linear search.
