+++ z/btree_test.go
@@ -19,12 +19,13 @@ package z
 import (
 	"fmt"
 	"io/ioutil"
+	"math"
 	"math/rand"
 	"os"
 	"sort"
 	"testing"
 
-	"github.com/dgraph-io/ristretto/contrib/simd"
+	"github.com/dgraph-io/ristretto/z/simd"
 	"github.com/dustin/go-humanize"
 	"github.com/stretchr/testify/require"
 )
@@ -314,7 +315,6 @@ func BenchmarkSearch(b *testing.B) {
 			return n.key(i) >= k
 		})
 	}
-
 	unroll4 := func(n node, k uint64, N int) int {
 		if len(n[:2*N]) < 8 {
 			for i := 0; i < N; i++ {
@@ -327,7 +327,8 @@ func BenchmarkSearch(b *testing.B) {
 		return int(simd.Search(n[:2*N], k))
 	}
 
-	for sz := 1; sz < 256; sz *= 2 {
+	jumpBy := []int{8, 16, 32, 64, 128, 196, 255}
+	for _, sz := range jumpBy {
 		f, err := ioutil.TempFile(".", "tree")
 		require.NoError(b, err)
 
@@ -343,7 +344,7 @@ func BenchmarkSearch(b *testing.B) {
 
 		b.Run(fmt.Sprintf("linear-%d", sz), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				tmp = linear(n, uint64(sz), sz)
+				tmp = linear(n, math.MaxUint64, sz)
 			}
 		})
 		b.Run(fmt.Sprintf("binary-%d", sz), func(b *testing.B) {
@@ -353,7 +354,7 @@ func BenchmarkSearch(b *testing.B) {
 		})
 		b.Run(fmt.Sprintf("unrolled-asm-%d", sz), func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				tmp = unroll4(n, uint64(sz), sz)
+				tmp = unroll4(n, math.MaxUint64, sz)
 			}
 		})
 		mf.Close(0)
