+++ z/simd/baseline.go
