+++ z/simd/search.s
