+++ z/simd/stub_search.go
